몬스터 체력 표시 UI 설정

+++ Assets/Boss.cs
@@ -32,8 +32,8 @@ public class Boss : MonoBehaviour
 
     private bool Damaged = false;
 
-    public int curHealth;
-    public int maxHealth;
+    public float curHealth;
+    public float maxHealth;
 
     public GameObject[] items;
     public BoxCollider meleeArea;
